requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
grpcio>=1.50.0
grpcio-tools>=1.50.0
pyyaml>=6.0
numpy>=1.24.0
pydantic>=2.0.0
ruff>=0.1.0
pre-commit>=3.5.0
pytest>=7.4.0
pytest-cov>=4.1.0
onnxruntime>=1.16.0
onnxscript>=0.1.0
optimum[onnxruntime]>=1.14.0
psutil>=5.9.0
playwright>=1.40.0
tqdm>=4.65.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Apple Silicon native ML framework (optional, for INT4/INT8 quantization)
# Only works on Apple Silicon Macs
# mlx>=0.0.10
# mlx-lm>=0.0.10
